Add a lab with JdbcTemplate
This is just a small lab to demonstrate how to use JdbcTemplate instead
of JPA. The leverage with JdbcTemplate over JPA is that you don't need
to have any Entities. You can just map the database results directly to
your domain objects.

Note that the downside of this is that you need to manually map
one-to-many relationships by hand when creating/mapping your domain
objects.